requirements.txt
pandas>=2.0
numpy>=1.24
tqdm>=4.66

# Deep learning
torch>=2.0
mlflow>=2.14

# PyG (torch-geometric) installation can be platform-specific.
# If pip install fails, follow the official wheel instructions for your torch/cuda version.
torch-geometric

# Chemistry
# If rdkit-pypi fails on your platform, install RDKit via conda-forge instead.
rdkit-pypi
selfies>=2.1

# CID -> SMILES lookup
PubChemPy>=1.0
